+++ internal/agent/services/data_store/strategies/otlp/metric_store.go
@@ -166,7 +166,7 @@ func (s *metricStore) upsert(dp datapoint.DataPoint) {
 		probeName:  probeName,
 		probeType:  probeType,
 		metricName: dp.Name,
-		value:      float64(dp.Value),
+		value:      dp.Value,
 		unit:       tagMap["unit"],
 		tags:       tagsCopy,
 		observedAt: when,
